Bluetooth: Add Manifest configure to fix settings crash issue.
Settings is crashed when press back key after rotating screen when
Bluetooth profile activity is opened. Add manifest configure to
enable screen orientation feature.

Settings: Change the text size of title when rotate screen

The text size of bluetooth settings' title is not changed when
rotate the screen.
According to the configuration, change the text size of title.

Change-Id: I3f08d0dd8181ffc8999ed29ac31dcc19a7dd56e9
CRs-Fixed: 917010