Settings: Enable storage manager
This feature was introduced in 7.1 but turned on only on the 'Pixel' device. I've
used it on the angler and shamu for a few days now and can't find any jank.

Anyone with questions can read more about it here.
http://www.androidauthority.com/android-7-1-nougat-smart-storage-manager-723501/

Credits to @snak3ater for finding it
https://github.com/nitrogen-project/android_packages_apps_Settings/commit/dd601552a4ab13baee439fa4c72a8c6669247503
https://github.com/nitrogen-project/android_vendor_nitrogen/commit/9f23de2110a8bd936e2e7ad4791d3b372b2fac53

The problem with the above (at least for me) is that the bool was not doing much of anything and inheriting a
system property wasn't something I wanted to do.

Revert "Show storage manager settings using a system property."

This reverts commit 00748552297fe45ddf55878b3f37c79c1863afa2.

Change-Id: If975137e8bfcd5c5448e12a97bb897fd3011360d