Custom app dashboard icons: Update coloring location
With this commit: http://review.projektsubstratum.com/#/c/259/,
we need to make sure that the tint is applied like all of the
other dashboard icons.

Without this commit, if a themer decides to take advantage of
the fact they can use a seperate color for the settings icons,
these icons will miss being tinted and will take on the accent.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>